colorLinkの修正
