--- src/components/colorLink/colorLink.module.css
@@ -1,9 +1,0 @@
-.LinkText {
-  color: var(--blueColor);
-  text-decoration: underline;
-  font-size: 1.1rem;
-}
-
-.LinkText:hover {
-  font-weight: bold;
-}
