+++ src/pages/index.tsx
@@ -1,10 +1,17 @@
-import Dialog from "@/components/dialog/Dialog";
-import EditRoundFrame from "@/components/editRoundFrame/EditRoundFrame";
 import Layout from "@/components/layout/Layout";
-import MarkDown from "@/components/markDown/MarkDown";
-import RoundFrame from "@/components/roundFrame/RoundFrame";
+import Lists from "@/components/lists/Lists";
+import useSWR from "swr";
+
+const fetcher = (url: string) => fetch(url).then((res) => res.json());
 
 export default function Home() {
+  const { data, error, isLoading } = useSWR(
+    "http://localhost:8000/articles",
+    fetcher
+  );
+
+  if (error) return <div>failed to load</div>;
+  if (isLoading) return <div>loading...</div>;
   return (
     <>
       <Layout
@@ -13,18 +20,7 @@ export default function Home() {
         headTitle="Raha"
         pageTitle="投稿一覧"
       >
-        <EditRoundFrame>
-          <p>作成日：2025-1-7</p>
-          <p>作成者：山竹森楓奏</p>
-          <p>「javaの基礎をまとめました」</p>
-          <p>#はじめに</p>
-          <p>　　- Java初学者です</p>
-          <p>#Javaの基礎</p>
-          <p>　　- Java初学者です</p>
-        </EditRoundFrame>{" "}
-
-        <MarkDown buttonText="投稿" textarea="" title=""></MarkDown>
-
+        <Lists pagedata={data}></Lists>
       </Layout>
     </>
   );
